+++ app/wyze_bridge.py
@@ -7,7 +7,7 @@ from typing import NoReturn
 from wyzebridge import config
 from wyzebridge.bridge_utils import env_bool, env_cam
 from wyzebridge.logging import logger
-from wyzebridge.rtsp_server import RtspServer
+from wyzebridge.rtsp_server import MtxServer
 from wyzebridge.stream import StreamManager
 from wyzebridge.wyze_api import WyzeApi
 from wyzebridge.wyze_stream import WyzeStream, WyzeStreamOptions
@@ -23,7 +23,7 @@ class WyzeBridge(Thread):
         print(f"\n🚀 STARTING DOCKER-WYZE-BRIDGE v{config.VERSION}\n")
         self.api: WyzeApi = WyzeApi()
         self.streams: StreamManager = StreamManager()
-        self.rtsp: RtspServer = RtspServer(config.BRIDGE_IP)
+        self.rtsp: MtxServer = MtxServer(config.BRIDGE_IP)
 
         if config.LLHLS:
             self.rtsp.setup_llhls(config.TOKEN_PATH, bool(config.HASS_TOKEN))
@@ -31,10 +31,10 @@ class WyzeBridge(Thread):
     def run(self, fresh_data: bool = False) -> None:
         self.api.login(fresh_data=fresh_data)
         self.setup_streams()
-        self.rtsp.start()
         if self.streams.total < 1:
-            return self.clean_up()
-        self.streams.monitor_streams()
+            return signal.raise_signal(signal.SIGINT)
+        self.rtsp.start()
+        self.streams.monitor_streams(self.rtsp.health_check)
 
     def setup_streams(self):
         """Gather and setup streams for each camera."""
@@ -42,17 +42,18 @@ class WyzeBridge(Thread):
         WyzeStream.api = self.api
         for cam in self.api.filtered_cams():
             logger.info(f"[+] Adding {cam.nickname} [{cam.product_model}]")
-            if config.SNAPSHOT_TYPE == "API":
+            if config.SNAPSHOT_TYPE == "api":
                 self.api.save_thumbnail(cam.name_uri)
             options = WyzeStreamOptions(
                 quality=env_cam("quality", cam.name_uri),
                 audio=bool(env_cam("enable_audio", cam.name_uri)),
                 record=bool(env_cam("record", cam.name_uri)),
+                reconnect=not config.ON_DEMAND,
             )
             self.add_substream(cam, options)
             stream = WyzeStream(cam, options)
             stream.rtsp_fw_enabled = self.rtsp_fw_proxy(cam, stream)
-            self.rtsp.add_path(stream.uri, not bool(options.record))
+            self.rtsp.add_path(stream.uri, not options.reconnect)
             self.streams.add(stream)
 
     def rtsp_fw_proxy(self, cam, stream) -> bool:
@@ -69,9 +70,10 @@ class WyzeBridge(Thread):
         if env_bool(f"SUBSTREAM_{cam.name_uri}") or (
             env_bool("SUBSTREAM") and cam.can_substream
         ):
-            sub_opt = replace(options, quality="sd30", substream=True)
+            quality = env_bool("sub_quality", "sd30")
+            sub_opt = replace(options, quality=quality, substream=True)
             sub = WyzeStream(cam, sub_opt)
-            self.rtsp.add_path(sub.uri, on_demand=True)
+            self.rtsp.add_path(sub.uri, not options.reconnect)
             self.streams.add(sub)
 
     def clean_up(self) -> NoReturn: