+++ app/wyzebridge/wyze_stream.py
@@ -42,9 +42,14 @@ class WyzeStreamOptions:
     quality: str = "hd180"
     audio: bool = False
     record: bool = False
+    reconnect: bool = False
     substream: bool = False
     frame_size: int = 0
     bitrate: int = 120
+
+    def __post_init__(self):
+        if self.record:
+            self.reconnect = True
 
     def update_quality(self, is_2k: bool = False) -> None:
         quality = (self.quality or "na").lower().ljust(3, "0")
@@ -177,7 +182,9 @@ class WyzeStream:
                 self.start_time = time() + COOLDOWN
                 logger.info(f"{self.camera.nickname} will cooldown for {COOLDOWN}s.")
         elif (
-            self.state == StreamStatus.STOPPED and self.options.record and should_start
+            self.state == StreamStatus.STOPPED
+            and self.options.reconnect
+            and should_start
         ):
             self.start()
         elif self.state == StreamStatus.CONNECTING and is_timedout(self.start_time, 20):
@@ -206,7 +213,7 @@ class WyzeStream:
             "status": self.state,
             "connected": self.connected,
             "enabled": self.enabled,
-            "on_demand": not self.options.record,
+            "on_demand": not self.options.reconnect,
             "audio": self.options.audio,
             "record": self.options.record,
             "substream": self.options.substream,